Reorganize and prevent double loading
